create file for batch move_dir_content and  move data_management functions and correct loops within merge
